Don't require ILspMiscellaneousFilesWorkspace to be a workspace
This coupling can cause challenges, so the easy change is to add a
Workspace property which provides the workspace. This renames the
ILspMiscellaneousFilesWorkspace interface to add a Provider suffix,
and the existing Provider now becomes a ProviderFactory.
