+++ src/LanguageServer/Protocol/RoslynLanguageServer.cs
@@ -98,7 +98,7 @@ internal sealed class RoslynLanguageServer : SystemTextJsonLanguageServer<Reques
         // those cases, we do not need to add an additional workspace to manage new files we hear about.  So only
         // add the LspMiscellaneousFilesWorkspace for hosts that have not already brought their own.
         if (serverKind == WellKnownLspServerKinds.CSharpVisualBasicLspServer)
-            AddLazyService<ILspMiscellaneousFilesWorkspace>(lspServices => lspServices.GetRequiredService<ILspMiscellaneousFilesWorkspaceProvider>().CreateLspMiscellaneousFilesWorkspace(lspServices, hostServices));
+            AddLazyService<ILspMiscellaneousFilesWorkspaceProvider>(lspServices => lspServices.GetRequiredService<ILspMiscellaneousFilesWorkspaceProviderFactory>().CreateLspMiscellaneousFilesWorkspaceProvider(lspServices, hostServices));
 
         return baseServiceMap.ToFrozenDictionary(
             keySelector: kvp => kvp.Key,
