+++ src/LanguageServer/Protocol/Workspaces/LspWorkspaceManager.cs
@@ -115,19 +115,19 @@ internal sealed class LspWorkspaceManager : IDocumentChangeTracker, ILspService
     private ImmutableDictionary<Uri, (SourceText Text, string LanguageId)> _trackedDocuments = ImmutableDictionary<Uri, (SourceText, string)>.Empty.WithComparers(LspUriComparer.Instance);
 
     private readonly ILspLogger _logger;
-    private readonly ILspMiscellaneousFilesWorkspace? _lspMiscellaneousFilesWorkspace;
+    private readonly ILspMiscellaneousFilesWorkspaceProvider? _lspMiscellaneousFilesWorkspaceProvider;
     private readonly LspWorkspaceRegistrationService _lspWorkspaceRegistrationService;
     private readonly ILanguageInfoProvider _languageInfoProvider;
     private readonly RequestTelemetryLogger _requestTelemetryLogger;
 
     public LspWorkspaceManager(
         ILspLogger logger,
-        ILspMiscellaneousFilesWorkspace? lspMiscellaneousFilesWorkspace,
+        ILspMiscellaneousFilesWorkspaceProvider? lspMiscellaneousFilesWorkspace,
         LspWorkspaceRegistrationService lspWorkspaceRegistrationService,
         ILanguageInfoProvider languageInfoProvider,
         RequestTelemetryLogger requestTelemetryLogger)
     {
-        _lspMiscellaneousFilesWorkspace = lspMiscellaneousFilesWorkspace;
+        _lspMiscellaneousFilesWorkspaceProvider = lspMiscellaneousFilesWorkspace;
         _logger = logger;
         _requestTelemetryLogger = requestTelemetryLogger;
 
@@ -197,7 +197,7 @@ internal sealed class LspWorkspaceManager : IDocumentChangeTracker, ILspService
         _cachedLspSolutions.Clear();
 
         // Also remove it from our loose files or metadata workspace if it is still there.
-        _lspMiscellaneousFilesWorkspace?.TryRemoveMiscellaneousDocument(uri, removeFromMetadataWorkspace: true);
+        _lspMiscellaneousFilesWorkspaceProvider?.TryRemoveMiscellaneousDocument(uri, removeFromMetadataWorkspace: true);
 
         LspTextChanged?.Invoke(this, EventArgs.Empty);
 
@@ -295,10 +295,10 @@ internal sealed class LspWorkspaceManager : IDocumentChangeTracker, ILspService
 
                 // As we found the document in a non-misc workspace, also attempt to remove it from the misc workspace
                 // if it happens to be in there as well.
-                if (workspace != _lspMiscellaneousFilesWorkspace)
+                if (workspace != _lspMiscellaneousFilesWorkspaceProvider?.Workspace)
                 {
                     // Do not attempt to remove the file from the metadata workspace (the document is still open).
-                    _lspMiscellaneousFilesWorkspace?.TryRemoveMiscellaneousDocument(uri, removeFromMetadataWorkspace: false);
+                    _lspMiscellaneousFilesWorkspaceProvider?.TryRemoveMiscellaneousDocument(uri, removeFromMetadataWorkspace: false);
                 }
 
                 return (workspace, document.Project.Solution, document);
@@ -311,10 +311,10 @@ internal sealed class LspWorkspaceManager : IDocumentChangeTracker, ILspService
         _logger.LogInformation($"Could not find '{textDocumentIdentifier.Uri}'.  Searched {searchedWorkspaceKinds}");
         _requestTelemetryLogger.UpdateFindDocumentTelemetryData(success: false, workspaceKind: null);
 
-        // Add the document to our loose files workspace (if we have one) if it iss open.
+        // Add the document to our loose files workspace (if we have one) if it is open.
         if (_trackedDocuments.TryGetValue(uri, out var trackedDocument))
         {
-            var miscDocument = _lspMiscellaneousFilesWorkspace?.AddMiscellaneousDocument(uri, trackedDocument.Text, trackedDocument.LanguageId, _logger);
+            var miscDocument = _lspMiscellaneousFilesWorkspaceProvider?.AddMiscellaneousDocument(uri, trackedDocument.Text, trackedDocument.LanguageId, _logger);
             if (miscDocument is not null)
                 return (miscDocument.Project.Solution.Workspace, miscDocument.Project.Solution, miscDocument);
         }
@@ -584,10 +584,7 @@ internal sealed class LspWorkspaceManager : IDocumentChangeTracker, ILspService
 
         public Workspace? GetLspMiscellaneousFilesWorkspace()
         {
-            // Right now this assumes for purposes of tests that the implementation of ILspMiscellaneousFilesWorkspace is itself a Workspace.
-            // If we wanted to decouple this better, we could add a member to ILspMiscellaneousFilesWorkspace, but since this is only for tests
-            // it's not the worst.
-            return (Workspace?)_manager._lspMiscellaneousFilesWorkspace;
+            return _manager._lspMiscellaneousFilesWorkspaceProvider?.Workspace;
         }
 
         public bool IsWorkspaceRegistered(Workspace workspace)
