+++ src/VisualStudio/Core/Def/LanguageClient/VisualStudioLspWorkspaceRegistrationService.cs
@@ -16,17 +16,4 @@ internal sealed class VisualStudioLspWorkspaceRegistrationService : LspWorkspace
     public VisualStudioLspWorkspaceRegistrationService()
     {
     }
-
-    public override void Register(Workspace? workspace)
-    {
-        // The lsp misc files workspace has the MiscellaneousFiles workspace kind,
-        // but we don't actually want to mark it as a registered workspace in VS since we
-        // prefer the actual MiscellaneousFilesWorkspace.
-        if (workspace is LspMiscellaneousFilesWorkspace)
-        {
-            return;
-        }
-
-        base.Register(workspace);
-    }
 }
